chore: update minimum supported react version in package.json
