+++ package.json
@@ -77,8 +77,8 @@
     "prettier": "^2.4.1",
     "process": "^0.11.10",
     "prop-types": "^15.7.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "16.14.0",
+    "react-dom": "16.14.0",
     "react-test-renderer": "^18.2.0",
     "rimraf": "^3.0.2",
     "rollup": "2.79.1",
@@ -102,8 +102,8 @@
     "typescript": "^4.8.4"
   },
   "peerDependencies": {
-    "react": ">=18.0.0",
-    "react-dom": ">=18.0.0"
+    "react": ">=16.14.0",
+    "react-dom": ">=16.14.0"
   },
   "browserslist": {
     "production": [
