+++ src/index-dev.tsx
@@ -1,16 +1,8 @@
-import { StrictMode, version } from 'react'
-import { createRoot } from 'react-dom/client'
-import './tokens.css'
+import React, { version } from 'react'
+import ReactDOM from 'react-dom'
 import App from './App'
+import './tokens.css'
 
-// eslint-disable-next-line no-console
 console.log('Parent folder loaded react version: ', version)
 
-const container = document.getElementById('app')
-// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-const root = createRoot(container!)
-root.render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-)
+ReactDOM.render(<App />, document.getElementById('app'))
